requirements.txt
python-twitter
BeautifulSoup
google-api-python-client
jira
python-wordpress-xmlrpc
slacker